Rename *Observable into *Observer
