+++ firebase-auth/src/main/java/com/androidhuman/rxfirebase2/auth/SignInAnonymouslyObserver.java
@@ -1,50 +1,49 @@
 package com.androidhuman.rxfirebase2.auth;
 
 import com.google.android.gms.tasks.Task;
+import com.google.firebase.auth.AuthResult;
 import com.google.firebase.auth.FirebaseAuth;
+import com.google.firebase.auth.FirebaseUser;
 
 import com.androidhuman.rxfirebase2.core.OnCompleteDisposable;
 
 import android.support.annotation.NonNull;
 
-import io.reactivex.Completable;
-import io.reactivex.CompletableObserver;
+import io.reactivex.Single;
+import io.reactivex.SingleObserver;
 
-final class SendPasswordResetEmailObservable extends Completable {
+final class SignInAnonymouslyObserver extends Single<FirebaseUser> {
 
     private final FirebaseAuth instance;
 
-    private final String email;
-
-    SendPasswordResetEmailObservable(FirebaseAuth instance, String email) {
+    SignInAnonymouslyObserver(FirebaseAuth instance) {
         this.instance = instance;
-        this.email = email;
     }
 
     @Override
-    protected void subscribeActual(CompletableObserver observer) {
+    protected void subscribeActual(SingleObserver<? super FirebaseUser> observer) {
         Listener listener = new Listener(observer);
         observer.onSubscribe(listener);
 
-        instance.sendPasswordResetEmail(email)
+        instance.signInAnonymously()
                 .addOnCompleteListener(listener);
     }
 
-    static final class Listener extends OnCompleteDisposable<Void> {
+    static final class Listener extends OnCompleteDisposable<AuthResult> {
 
-        private final CompletableObserver observer;
+        private final SingleObserver<? super FirebaseUser> observer;
 
-        Listener(@NonNull CompletableObserver observer) {
+        Listener(@NonNull SingleObserver<? super FirebaseUser> observer) {
             this.observer = observer;
         }
 
         @Override
-        public void onComplete(@NonNull Task<Void> task) {
+        public void onComplete(@NonNull Task<AuthResult> task) {
             if (!isDisposed()) {
                 if (!task.isSuccessful()) {
                     observer.onError(task.getException());
                 } else {
-                    observer.onComplete();
+                    observer.onSuccess(task.getResult().getUser());
                 }
             }
         }
