--- firebase-auth/src/main/java/com/androidhuman/rxfirebase2/auth/SignInWithCustomTokenObservable.java
@@ -1,54 +1,0 @@
-package com.androidhuman.rxfirebase2.auth;
-
-import com.google.android.gms.tasks.Task;
-import com.google.firebase.auth.AuthResult;
-import com.google.firebase.auth.FirebaseAuth;
-import com.google.firebase.auth.FirebaseUser;
-
-import com.androidhuman.rxfirebase2.core.OnCompleteDisposable;
-
-import android.support.annotation.NonNull;
-
-import io.reactivex.Single;
-import io.reactivex.SingleObserver;
-
-final class SignInWithCustomTokenObservable extends Single<FirebaseUser> {
-
-    private final FirebaseAuth instance;
-
-    private final String token;
-
-    SignInWithCustomTokenObservable(FirebaseAuth instance, String token) {
-        this.instance = instance;
-        this.token = token;
-    }
-
-    @Override
-    protected void subscribeActual(SingleObserver<? super FirebaseUser> observer) {
-        Listener listener = new Listener(observer);
-        observer.onSubscribe(listener);
-
-        instance.signInWithCustomToken(token)
-                .addOnCompleteListener(listener);
-    }
-
-    static final class Listener extends OnCompleteDisposable<AuthResult> {
-
-        private final SingleObserver<? super FirebaseUser> observer;
-
-        Listener(@NonNull SingleObserver<? super FirebaseUser> observer) {
-            this.observer = observer;
-        }
-
-        @Override
-        public void onComplete(@NonNull Task<AuthResult> task) {
-            if (!isDisposed()) {
-                if (!task.isSuccessful()) {
-                    observer.onError(task.getException());
-                } else {
-                    observer.onSuccess(task.getResult().getUser());
-                }
-            }
-        }
-    }
-}
